# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Weynon_V2",
    "description": "Welcome to Weynon Api",
    "contact": {}
  },
  "host": "178.79.177.139:8848/api",
  "basePath": "/",
  "securityDefinitions": {
    "auth": {
      "type": "oauth2",
      "flow": "implicit",
      "authorizationUrl": "http://178.79.177.139:8848/api",
      "scopes": {}
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users": {
      "get": {
        "description": "TODO: Add Description",
        "summary": "Get All User in the system",
        "tags": [
          "Users"
        ],
        "operationId": "UsersGet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/users/get-contacts-detail": {
      "post": {
        "description": "TODO: Add Description",
        "summary": "Contacts Detail",
        "tags": [
          "Users"
        ],
        "operationId": "UsersGetContactsDetailPost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ContactsDetailRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
...
    "/friends": {
      "get": {
        "description": "TODO: Add Description",
        "summary": "Get All Friends",
        "tags": [
          "Friends"
        ],
        "operationId": "FriendsGet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "ContactsDetailRequest": {
      "title": "ContactsDetailRequest",
      "example": {
        "contacts": [
          {
            "name": "Muhammad Asif javed",
            "phonenumber": "+923324647331"
          },
          {
            "name": "Muhammad Asif javed",
            "phonenumber": "+923320000000"
          },
          {
            "name": "Muhammad Kashif javed",
            "phonenumber": "+923324647482"
          },
          {
            "name": "Ajlal Khan",
            "phonenumber": "+923170770556"
          },
          {
            "name": "Foo",
            "phonenumber": "+923324647331"
          },
          {
            "name": "bar",
            "phonenumber": "+923320000000"
          },
          {
            "name": "bazzz",
            "phonenumber": "+923000000000"
          },
          {
            "name": "",
            "phonenumber": "+923324647332"
          }
        ]
      },
      "type": "object",
      "properties": {
        "contacts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Contact"
          }
        }
      },
      "required": [
        "contacts"
      ]
    },
    "Contact": {
      "title": "Contact",
      "type": "object",
      "properties": {
        "name": {
          "example": "Muhammad Asif javed",
          "type": "string"
        },
        "phonenumber": {
          "example": "+923324647331",
          "type": "string"
        }
      },
      "required": [
        "name",
        "phonenumber"
      ]
    },
...
  },
...
}